Switched int with long

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        int[] salesOfManager1 = {20, 23, 10, 3, 35};
+        long[] salesOfManager1 = {20, 23, 10, 3, 35};
         SalesManager manager1 = new SalesManager(salesOfManager1);
 
         System.out.println("Максимальные продажи Менеджера №1 за текущую неделю составили: " + manager1.max() + ".");
